reconciler/workload/resource+namespace: drop code for other side, and fixup
Should be functionally equivalent.
